DanceParty will load p5 itself, play() takes songData param
